Fix speakerphone in JB 4.1 (1/2)
On Android 4.1 the speakerphone does not work correctly - the volume remains low and the audio server often crashes and causes a hot reboot of the device.
The problem is mainly caused by the p500 audio libraries which are not needed and just give problems. This first part of the commit removes those libraries.

Change-Id: I330fb864f52612a66637ba962be3e0e681fffd22